Show composition notes on view page
