+++ app/assets/stylesheets/composition-view.scss
@@ -15,3 +15,11 @@
   line-height: 1.3;
   margin-bottom: 5px;
 }
+
+.players-view-table.players-table {
+  tbody td {
+    padding-left: 0.75em;
+    padding-right: 0.75em;
+    line-height: 1.5;
+  }
+}
