+++ app/assets/stylesheets/forms.scss
@@ -103,7 +103,6 @@
     height: 2.285em;
     line-height: 1.5;
     padding-left: 0.75em;
-    padding-right: 0.75em;
     position: relative;
     vertical-align: top;
     background-color: inherit;
